requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
isort==5.13.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
